LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/layout-canvas-children-expected.txt


FAIL canvas children have layout assert_equals: expected 100 but got 0
